GIVE ME DAMN POPCICLE STICK

+++ Repositories/KeepsRepository.cs
@@ -40,12 +40,21 @@ namespace FindrsKeeprs.Repositories
         throw new Exception("Keeps not found");
       }
     }
-    public Keep GetKeepById(int Id)
+    public Keep GetKeepById(int id, int isprivate)
     {
       try
       {
+        //FIXME dont allow users to get keeps by id if the keep is private
+        //NOTE only allowed isPrivate=0 keeps to be accessed
+        var keep = _db.QueryFirstOrDefault<Keep>("SELECT * FROM keeps WHERE id=@Id AND isPrivate=@IsPrivate", new { id, isprivate });
+        if (keep != null)
+        {
+          //TODO UPDATE THE KEEP VIEWS
+          keep.Views++;
 
-        return _db.QueryFirstOrDefault<Keep>("SELECT * FROM keeps WHERE id=@Id", new { Id });
+
+        }
+        return keep;
       }
       catch (Exception e)
       {
@@ -53,9 +62,11 @@ namespace FindrsKeeprs.Repositories
       }
     }
 
-    public bool DeleteKeepById(int id)
+    public bool DeleteKeepById(int id, string userId)
     {
-      int success = _db.Execute("DELETE FROM keeps WHERE id=@Id", new { id });
+      //FIXME only the creator of the keep should be able to delete the keep
+      //NOTE added userId
+      int success = _db.Execute("DELETE FROM keeps WHERE id=@Id AND userId=@UserID", new { id, userId });
       return success > 0;
     }
   }
